Extract logging functionality into dedicated logging.rs module, and clean main.rs of logging infrastructure.
- Move logging code from main.rs and file_io.rs to new logging.rs module
- Extract crash logging, log exports, stdout capture, and signal handlers
- Update all imports and function calls across codebase
