+++ src/app.rs
@@ -792,14 +792,14 @@ impl iced_winit::runtime::Program for DataViewer {
             }
             Message::ShowLogs => {
                 let app_name = "viewskater";
-                let log_dir_path = file_io::get_log_directory(app_name);
+                let log_dir_path = crate::logging::get_log_directory(app_name);
                 let _ = std::fs::create_dir_all(log_dir_path.clone());
-                file_io::open_in_file_explorer(&log_dir_path.to_string_lossy().to_string());
+                crate::logging::open_in_file_explorer(&log_dir_path.to_string_lossy().to_string());
             }
             Message::ExportDebugLogs => {
                 let app_name = "viewskater";
                 if let Some(log_buffer) = crate::get_shared_log_buffer() {
-                    file_io::export_and_open_debug_logs(app_name, log_buffer);
+                    crate::logging::export_and_open_debug_logs(app_name, log_buffer);
                 } else {
                     warn!("Log buffer not available for export");
                 }
@@ -811,11 +811,11 @@ impl iced_winit::runtime::Program for DataViewer {
                     println!("DEBUG: Got log buffer, starting export...");
                     if let Some(stdout_buffer) = crate::get_shared_stdout_buffer() {
                         println!("DEBUG: Got stdout buffer, calling export_and_open_all_logs...");
-                        file_io::export_and_open_all_logs(app_name, log_buffer, stdout_buffer);
+                        crate::logging::export_and_open_all_logs(app_name, log_buffer, stdout_buffer);
                         println!("DEBUG: export_and_open_all_logs completed");
                     } else {
                         println!("DEBUG: Stdout buffer not available, exporting debug logs only");
-                        match file_io::export_debug_logs(app_name, log_buffer) {
+                        match crate::logging::export_debug_logs(app_name, log_buffer) {
                             Ok(debug_log_path) => {
                                 println!("DEBUG: Export successful to: {}", debug_log_path.display());
                                 info!("Debug logs successfully exported to: {}", debug_log_path.display());
